Added initial vendored dependencies
;; WARNING: huge commit, roughly 15 MB of source files!

;; About

Due to instability in the crates.io ecosystem, the
dependencies of azul have to be currently vendored
in order to provide stability and reproducible builds.

The goal of this repository is to optimize the
dependencies as far as possible and optimize this repo
specifically for azuls use-case. This means that there
will be (documented) changes to the source files.

The final crate will / should re-export all crates
listed here, so that azul (a) only uses one version
of any given crate, (b) reliably builds and (c) isn't
victim to random crates yanking their code from cargo.

Vendoring sources was pretty much inevitable anyways,
since there are azul-specific changes to certain repositories.
For example, if azul doesn't use a specific feature or branch,
the entire dependency will be simply removed, instead of
the user having to download and compile optional crates.

Optimizations that need to be done:

- winit currently has an entire CPU backend, which can render GL
  output to the terminal via libcaca. While it's a cool gimmick, it adds
  nothing to the usability, and libcaca should be nuked from the
  dependencies.
- webrender currently depends on pathfinder, but pathfinder is way
  too immature to be used in production. Therefore, pathfinder will
  be nuked from the dependencies.
- winit currently depends on the entire parking-lot ecosystem, just
  to make a single mutex that is only used a few times faster and to
  make winit compile on Win32. Since azul itself uses std::sync::Mutex
  and is therefore not portable to Windows XP, this is fairly useless.
  parking-lot will be nuked, since it is a huge dependency.
- rand, freetype, xml-rs, etc. are duplicated dependencies that should
  be manually merged (because cargo was too stupid to do so).
- In the initial commit, the servo-freetype-sys/doc folder was removed
  because it contained a lot of small HTML files, that aren't source files
  and are therefore not relevant to actually building freetype.
- webrender has a lot of profiling infrastructure, that can be removed
  for azul (since azul does not use the profiling infrastructure).
- webrender currently depends on serde, which is both a compile-time
  and a runtime cost. In the end, serde should be removed from the
  dependencies.

crates.io does not allow crates larger than 5MB source code, so the
goal is to get below that line. Further optimizations to removing
unnecessary feature-flags and therefore unnecessary crates will have
to be done in the future.

;; Licensing

All sub-crates have their own LICENSE file, but are compatible with
the MIT license. This means that the global LICENSE for all dependencies
is MIT, not Apache, since MIT is more permissive.
Further licensing information with proper name attribution should
be added in the future, this is a work in progress.